Syncing data on Parse cloud

+++ scripts/services/classes/entry.js
@@ -17,6 +17,12 @@ app.factory("Entry", ["$q", "$indexedDB", "entries", "Player", function getEntry
             this.age = entryData.age;
             this.properties = entryData.properties;
             this.player = Player.getById(entryData.playerId);
+
+            if (entryData.deleted)
+                this._deleted = entryData.deleted;
+
+            if (entryData.cloudId)
+                this.cloudId = entryData.cloudId;
         }
         else{
             this.date = new Date();
@@ -44,23 +50,47 @@ app.factory("Entry", ["$q", "$indexedDB", "entries", "Player", function getEntry
     }
 
     Entry.prototype = {
-        remove: function () {
+        isNewEntry: function(){
+            return !this.timestamp;
+        },
+        getCloudData: function(){
+            return {
+                playerId: this.player.playerId,
+                age: this.age,
+                timestamp: this.timestamp,
+                date: this.date,
+                properties: this.properties,
+                type: this.type.id,
+                id: this.cloudId,
+                deleted: !!this._deleted
+            }
+        },
+        remove: function (absoluteDelete) {
             if (!this.timestamp)
                 throw new Error("Can't delete entry - it hasn't been saved yet.");
 
-            return entriesObjectStore.delete(this.timestamp).catch(function(error){
-                console.error("Can't delete entry: ", error);
-                return $q.reject("Can't delete entry");
-            });
+            if (absoluteDelete){
+                return entriesObjectStore.delete(this.timestamp).catch(function(error){
+                    console.error("Can't delete entry: ", error);
+                    return $q.reject("Can't delete entry");
+                });
+            }
+            else {
+                this._deleted = true;
+                this.save();
+            }
         },
-        save: function () {
+        save: function (isSynced) {
             if (!this.timestamp) {
                 this.isNewEntry = true;
                 this.timestamp = new Date().valueOf();
             }
-            else
+            else {
                 this.isNewEntry = false;
-
+                // The entry is deleted and the changed has been synced to cloud, can proceed to completely delete:
+                if (this._deleted && isSynced)
+                    return this.remove(true);
+            }
             var newEntry = this,
                 dbEntry = {
                     date: this.date,
@@ -69,8 +99,14 @@ app.factory("Entry", ["$q", "$indexedDB", "entries", "Player", function getEntry
                     type: this.type.id,
                     timestamp: this.timestamp,
                     playerId: this.player.playerId,
-                    synced: false
+                    cloudId: this.cloudId
                 };
+
+            if (!isSynced)
+                dbEntry.unsynced = 1;
+
+            if (this._deleted)
+                dbEntry.unsynced = dbEntry.deleted = 1;
 
             return entriesObjectStore.upsert(dbEntry).then(function (id) {
                 return newEntry;
@@ -84,13 +120,12 @@ app.factory("Entry", ["$q", "$indexedDB", "entries", "Player", function getEntry
         options = options || {};
 
         return entriesObjectStore.internalObjectStore(OBJECT_STORE_NAME, "readonly").then(function(objectStore){
-            //var objectStore = results[1];
-            var idx = objectStore.index(options.type ? "type_idx" : "date_idx");
+            var idx = objectStore.index(options.unsynced ? "unsync_idx" : options.type ? "type_idx" : "date_idx");
             var count = options.count || null,
                 entries = [],
                 currentRecord = 0,
                 deferred = $q.defer(),
-                cursorRange = IDBKeyRange.bound(
+                cursorRange = options.unsynced ? null : IDBKeyRange.bound(
                     options.type ? [options.playerId, options.type] : [options.playerId],
                     options.type ? [options.playerId, options.type, new Date()] : [options.playerId, new Date()]
                 ),
@@ -107,10 +142,12 @@ app.factory("Entry", ["$q", "$indexedDB", "entries", "Player", function getEntry
                     currentRecord = options.offset;
                     cursor.advance(options.offset);
                 }
-                else{
-                    entries.push(new Entry(cursor.value));
+                else {
+                    if(!cursor.value.deleted || options.includeDeleted) {
+                        entries.push(new Entry(cursor.value));
+                        currentRecord++;
+                    }
                     cursor.continue();
-                    currentRecord++;
                 }
             };
 
@@ -122,6 +159,10 @@ app.factory("Entry", ["$q", "$indexedDB", "entries", "Player", function getEntry
         }, function(){
             return $q.when([]);
         });
+    };
+
+    Entry.getUnsyncedEntries = function(){
+        return Entry.getEntries({ unsynced: true, includeDeleted: true });
     };
 
     return Entry;
